// src/modules/85-cv/components/CommonMultiItemsSideNav/components/CommonSelectedAppsSideNav/components/GroupedSideNav/GroupedSideNav.types.ts
/*
 * Copyright 2022 Harness Inc. All rights reserved.
 * Use of this source code is governed by the PolyForm Shield 1.0.0 license
 * that can be found in the licenses directory at the root of this repository, also available at
 * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
 */

import type { SelectOption } from '@harness/uicore'

export interface GroupedMetric {
  groupName?: SelectOption | string
  metricName?: string
  index?: number
  continuousVerification?: boolean
}

export interface GroupedCreatedMetrics {
  [Key: string]: GroupedMetric[]
}